Add 7.8.7 - Close step before the end of turn for priority window
The combat chain must be closed prior to the turn ending. The close step is outlined in 7.8 - specifically: 7.8.7. Fifth and finally, the Close Step ends, and the Action Phase continues.
The Action Phase will always continue after the combat chain is closed - so there is another round of priority windows