Earthlore bounty/effect context refactor
There was an issue with earthlore bounty/effect context. It was impossible to tell which player's class state the effect context would be in. That could lead to subtle bugs where the context could end up being missed. It was also a performance issue doing a lot of effect context setting and clearing on the decision queue IN EVERY GAME, even when neither player was using earthlore bounty. This should be much more efficient.